// src/app/blocks/block-types.ts
export enum BlockTypes {
  TEXT = 'TEXT',
  IMAGE = 'IMAGE',
  BLOG_V1 = 'BLOG_V1',
  BLOG_V2 = 'BLOG_V2',
  BLOG_V3 = 'BLOG_V3',
  BLOG_V4 = 'BLOG_V4',
  BLOG_V5 = 'BLOG_V5',
  CONTACT_V1 = 'CONTACT_V1',
  CONTENT_V1 = 'CONTENT_V1',
  CONTENT_V2 = 'CONTENT_V2',
  CONTENT_V3 = 'CONTENT_V3',
  CONTENT_V4 = 'CONTENT_V4',
  CONTENT_V5 = 'CONTENT_V5',
  CONTENT_V6 = 'CONTENT_V6',
  CONTENT_V7 = 'CONTENT_V7',
  FEATURE_V1 = 'FEATURE_V1',
  FEATURE_V2 = 'FEATURE_V2',
  FEATURE_V3 = 'FEATURE_V3',
  FEATURE_V4 = 'FEATURE_V4',
  FEATURE_V5 = 'FEATURE_V5',
  FEATURE_V6 = 'FEATURE_V6',
  GALLERY_V1 = 'GALLERY_V1',
  GALLERY_V2 = 'GALLERY_V2',
  GALLERY_V3 = 'GALLERY_V3',
  HERO_V1 = 'HERO_V1',
  HERO_V2 = 'HERO_V2',
  HERO_V3 = 'HERO_V3',
  HERO_IMAGE_FULL = 'HERO_IMAGE_FULL',
  STEP_V1 = 'STEP_V1',
  STEP_V2 = 'STEP_V2',
  TEAM_V1 = 'TEAM_V1',
  TEAM_V2 = 'TEAM_V2',
  TESTIMONIALS_V1 = 'TESTIMONIALS_V1',
  TESTIMONIALS_V2 = 'TESTIMONIALS_V2',
  TESTIMONIALS_V3 = 'TESTIMONIALS_V3',
  SOCIAL_V1 = 'SOCIAL_V1',
  CAROUSEL_V1 = 'CAROUSEL_V1'
}